DRY up missing_params
