+++ redash/utils/__init__.py
@@ -171,6 +171,23 @@ def collect_query_parameters(query):
     return keys
 
 
+def parameter_names(parameter_values):
+    names = []
+    for key, value in parameter_values.iteritems():
+        if isinstance(value, dict):
+            for inner_key in value.keys():
+                names.append(u'{}.{}'.format(key, inner_key))
+        else:
+            names.append(key)
+
+    return names
+
+
+def find_missing_params(query_text, parameter_values):
+    query_parameters = set(collect_query_parameters(query_text))
+    return set(query_parameters) - set(parameter_names(parameter_values))
+
+
 def collect_parameters_from_request(args):
     parameters = {}
